+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
-## [0.11.46] - 2020-07-06
+## [0.11.48] - 2020-07-06
 - Implemented Bar chart with token values
+
+## [0.11.47] - 2020-7-3
+- User is now transitioned to correct activity based on notification event_id
+
+## [0.11.46] - 2020-7-3
+- Added activity response completion on idle timer expiration
 
 ## [0.11.45] - 2020-6-23
 - Removed unnecessary fields and values in JSON-ld